fpgainfo: Add `add_dynamic_property` function
Add function `add_dynamic_property` that is used to add properties to
objects (rather that classes) depending on whether the relative sysfs
path exists. This should be used in the __init__ method of
sysfs_resource subclasses to add properties only if the relative path
exists.

This also changes `fme_errors` and `port_errors` to use this method.